Move handleOAuth
